+++ NEWS.md
@@ -5,7 +5,7 @@
   - The `ms_object` base class now has a `get_list_pager()` method which returns an object of class `ms_graph_pager`.
   - New `extract_list_values()` function to get all or part of the results from a paged result set.
 - The current (private) `ms_object$get_paged_list()` and `ms_object$init_list_objects()` methods are retained for backward compatibility, but are otherwise deprecated.
-- The `ms_graph$get_user()` method can now get a user by email or display name.
+- The `ms_graph$get_user()` method can now get a user by email or display name. Similarly, the `get_group()` method can get a group by display name.
 - Fix a bug in retrieving a paged list of values as a data frame, when `n` (the maximum number of rows) is supplied.
 - New `ms_graph$get_aad_object()` method to retrieve an Azure Active Directory object by ID. Mostly intended for use with the `list_object_memberships()` and `list_group_memberships()` methods, which return only IDs and not full object information.
 - All `list_*` R6 methods now have `filter` and `n` arguments to filter the result set and cap the number of results. The default values are `filter=NULL` and `n=Inf`. If `n=NULL`, the `ms_graph_pager` iterator object is returned instead to allow manual iteration over the results.